+++ src/cli-options.ts
@@ -104,6 +104,11 @@ export const optionDefinitions = [
     name: 'default-language',
     type: String,
     description: `The language used in the markerDefault funtion. Only available if use-marker-defaults is used.`
+  },
+  {
+    name: 'workdir',
+    type: String, 
+    description: `The workdir where the application is working from. Defaults to folder of execution.` 
   }
 ];